+++ packages/frameworks/src/frameworks.ts
@@ -2240,6 +2240,11 @@ export const frameworks = [
           path: 'Pipfile',
           matchContent: '[Dd]jango',
         },
+        {
+          // a default django project will create a manage.py which sets DJANGO_SETTINGS_MODULE
+          path: 'manage.py',
+          matchContent: 'DJANGO_SETTINGS_MODULE',
+        },
       ],
     },
     settings: {