+++ packages/backends/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vercel/backends
 
+## 0.0.40
+
+### Patch Changes
+
+- Updated dependencies [[`9721f4948a72d273f792c54f5393aa50ac279209`](https://github.com/vercel/vercel/commit/9721f4948a72d273f792c54f5393aa50ac279209)]:
+  - @vercel/build-utils@13.6.2
+
 ## 0.0.39
 
 ### Patch Changes