+++ packages/express/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vercel/express
 
+## 0.1.52
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.10
+  - @vercel/cervel@0.0.27
+
 ## 0.1.51
 
 ### Patch Changes